service/rag-service/requirements.txt
# FastAPI 서버
fastapi==0.110.0
uvicorn[standard]==0.29.0

# LangChain 및 벡터 저장소 (호환 버전)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.1.2
langchain-openai>=0.0.5

# 벡터 인덱스
faiss-cpu==1.7.4

# AI/ML 모델 (호환 버전으로 수정)
transformers>=4.37.0
torch>=2.1.2
accelerate>=0.25.0
tokenizers>=0.19.1
sentence-transformers>=2.2.0

# 데이터 처리 (NumPy 1.x로 고정하여 faiss 호환성 보장)
pandas>=2.1.4
numpy>=1.24.3,<2.0.0

# PDF 처리
PyPDF2>=3.0.1
pypdf>=3.0.0

# Excel 처리
openpyxl>=3.1.2

# 환경변수 및 설정
python-dotenv>=1.0.0
pydantic>=2.5.0
